+++ src/utils/convert_godot_lua.cpp
@@ -64,12 +64,7 @@ Variant to_variant(const sol::basic_object<ref_t>& object) {
 			return object.template as<double>();
 
 		case sol::type::table:
-			if (LuaScriptInstance *script_instance = LuaScriptInstance::find_instance(object.template as<sol::basic_table<ref_t>>())) {
-				return script_instance->owner;
-			}
-			else {
-				return LuaObject::wrap_object<LuaTable>(object);
-			}
+			return LuaObject::wrap_object<LuaTable>(object);
 
 		case sol::type::userdata:
 			if (object.template is<Variant>()) {
@@ -200,12 +195,6 @@ sol::stack_object lua_push(lua_State *lua_state, const Variant& value) {
 			if (LuaObject *lua_obj = Object::cast_to<LuaObject>(value)) {
 				if (LuaState::find_lua_state(lua_state) == lua_obj->get_lua_state().ptr()) {
 					sol::stack::push(lua_state, lua_obj->get_lua_object());
-					break;
-				}
-			}
-			if (LuaScriptInstance *script_instance = LuaScriptInstance::attached_to_object(value)) {
-				if (LuaState::find_lua_state(lua_state) == script_instance->get_lua_state().ptr()) {
-					sol::stack::push(lua_state, script_instance->data->get_table());
 					break;
 				}
 			}